requirements.txt
streamlit>=1.31.0
pandas>=2.0.0
python-dotenv>=1.0.0
google-genai>=1.0.0
rapidfuzz>=3.0.0
markdown_to_mrkdwn>=0.2.0
